Added link script and modified the makefile so that the memory test code is compiled and linked into a flat binary form, the same as for a bootsector.
It is way too big to fit in a bootsector, but I will probably test it by
making a bootsector that simply loads a preconfigured number of sectors
from immediately after the bootsector, ie no file system. The memory test
binary will then be placed onto the disk immediately after the bootsector.

memtest binary compiles and links as of this commit.