Fixed an issue with the bootsector where NUMSECTORS was not being pointed to the correct memory address.
To fix it, I changed the way NUMSECTORS was defined, to be defined in
start.s and mapped to an address by the link script, rather than being
defined in the link script.

As of this commit, the bootsector works, as tested with virtualbox, but
the second stage crashes horribly.