It Works!!
Added a subroutine to bios.s to read all of the memory descriptors into an
array, fixed bugs.

Changed output.c to use the new assembly subroutine to print a char.

Changed memtest.c so that all of the memory descriptors are kept in an
array rather than querying the BIOS for them one at a time.